+++ js/etherpad_nextcloud-embed-create-main.mjs
@@ -1,2 +1,2 @@
-import{o as q}from"./oc-compat-Cn8-bqPr.chunk.mjs";(function(){const n=document.getElementById("etherpad-nextcloud-embed-create");if(!(n instanceof HTMLElement))return;const c=Number(n.getAttribute("data-parent-folder-id")||""),d=String(n.getAttribute("data-create-by-parent-url")||"").trim(),w=String(n.getAttribute("data-request-token")||"").trim(),b=String(n.getAttribute("data-l10n-missing-name")||"Pad name is required."),f=String(n.getAttribute("data-l10n-invalid-access-mode")||"Invalid access mode."),h=String(n.getAttribute("data-l10n-incomplete-config")||"Embed configuration is incomplete."),m=n.querySelector("[data-epnc-embed-create-loading]"),l=n.querySelector("[data-epnc-embed-create-error]"),u=n.querySelector("[data-epnc-embed-create-error-message]"),g=()=>q(w),o=e=>{m instanceof HTMLElement&&(m.hidden=!0),u instanceof HTMLElement&&(u.textContent=String(e||"Unknown error.")),l instanceof HTMLElement&&(l.hidden=!1)},p=async(e,t={})=>{const i=new AbortController,r=window.setTimeout(()=>i.abort(),1e4),E=Object.assign({Accept:"application/json"},t.headers||{});try{const a=await fetch(e,Object.assign({credentials:"same-origin",headers:E,signal:i.signal},t)),s=await a.json().catch(()=>({}));if(!a.ok)throw new Error(s&&s.message||"Request failed.");return s}catch(a){throw a&&typeof a=="object"&&"name"in a&&a.name==="AbortError"?new Error("Request timed out."):a}finally{window.clearTimeout(r)}},S=()=>{const e=new URL(window.location.href).searchParams;return{name:String(e.get("name")||"").trim(),accessMode:String(e.get("accessMode")||"protected").trim()}},y=e=>{const t=new URL(String(e||"").trim(),window.location.origin);if(t.origin!==window.location.origin)throw new Error("Invalid embed URL origin.");return t.pathname+t.search+t.hash};(async()=>{if(!Number.isFinite(c)||c<=0||d===""){o(h);return}if(g()===""){o("CSRF request token is missing.");return}const{name:e,accessMode:t}=S();if(e===""){o(b);return}if(t!=="protected"&&t!=="public"){o(f);return}const i=new URLSearchParams;i.set("parentFolderId",String(c)),i.set("name",e),i.set("accessMode",t);try{const r=await p(d,{method:"POST",headers:{"Content-Type":"application/x-www-form-urlencoded;charset=UTF-8",requesttoken:g()},body:i.toString()});if(!r||typeof r.embed_url!="string"||r.embed_url.trim()==="")throw new Error("Pad creation API did not return a valid embed URL.");window.location.replace(y(r.embed_url))}catch(r){o(r instanceof Error?r.message:"Pad creation failed.")}})()})();
+import{o as q}from"./oc-compat-DDPKm7Cs.chunk.mjs";(function(){const n=document.getElementById("etherpad-nextcloud-embed-create");if(!(n instanceof HTMLElement))return;const c=Number(n.getAttribute("data-parent-folder-id")||""),d=String(n.getAttribute("data-create-by-parent-url")||"").trim(),w=String(n.getAttribute("data-request-token")||"").trim(),b=String(n.getAttribute("data-l10n-missing-name")||"Pad name is required."),f=String(n.getAttribute("data-l10n-invalid-access-mode")||"Invalid access mode."),h=String(n.getAttribute("data-l10n-incomplete-config")||"Embed configuration is incomplete."),m=n.querySelector("[data-epnc-embed-create-loading]"),l=n.querySelector("[data-epnc-embed-create-error]"),u=n.querySelector("[data-epnc-embed-create-error-message]"),g=()=>q(w),o=e=>{m instanceof HTMLElement&&(m.hidden=!0),u instanceof HTMLElement&&(u.textContent=String(e||"Unknown error.")),l instanceof HTMLElement&&(l.hidden=!1)},p=async(e,t={})=>{const i=new AbortController,r=window.setTimeout(()=>i.abort(),1e4),E=Object.assign({Accept:"application/json"},t.headers||{});try{const a=await fetch(e,Object.assign({credentials:"same-origin",headers:E,signal:i.signal},t)),s=await a.json().catch(()=>({}));if(!a.ok)throw new Error(s&&s.message||"Request failed.");return s}catch(a){throw a&&typeof a=="object"&&"name"in a&&a.name==="AbortError"?new Error("Request timed out."):a}finally{window.clearTimeout(r)}},S=()=>{const e=new URL(window.location.href).searchParams;return{name:String(e.get("name")||"").trim(),accessMode:String(e.get("accessMode")||"protected").trim()}},y=e=>{const t=new URL(String(e||"").trim(),window.location.origin);if(t.origin!==window.location.origin)throw new Error("Invalid embed URL origin.");return t.pathname+t.search+t.hash};(async()=>{if(!Number.isFinite(c)||c<=0||d===""){o(h);return}if(g()===""){o("CSRF request token is missing.");return}const{name:e,accessMode:t}=S();if(e===""){o(b);return}if(t!=="protected"&&t!=="public"){o(f);return}const i=new URLSearchParams;i.set("parentFolderId",String(c)),i.set("name",e),i.set("accessMode",t);try{const r=await p(d,{method:"POST",headers:{"Content-Type":"application/x-www-form-urlencoded;charset=UTF-8",requesttoken:g()},body:i.toString()});if(!r||typeof r.embed_url!="string"||r.embed_url.trim()==="")throw new Error("Pad creation API did not return a valid embed URL.");window.location.replace(y(r.embed_url))}catch(r){o(r instanceof Error?r.message:"Pad creation failed.")}})()})();
 //# sourceMappingURL=etherpad_nextcloud-embed-create-main.mjs.map
