--- js/urls-D-89WmZ3.chunk.mjs
@@ -1,2 +1,0 @@
-import{a as l,A as c}from"./oc-compat-Cn8-bqPr.chunk.mjs";const u=(n,t)=>{const e=!n||n==="/"?"":String(n),r=String(t||"").replace(/^\/+/,"").replace(/\s+\.pad$/i,".pad");return e===""?"/"+r:e+"/"+r},p=n=>typeof n=="string"&&n.toLowerCase().endsWith(".pad"),d=n=>{if(!n||typeof n!="string")return"/";const t=n.lastIndexOf("/");return t<=0?"/":n.slice(0,t)||"/"},f=()=>{const n=(new URLSearchParams(window.location.search||"").get("dir")||"/").trim();return n===""?"/":n},m=n=>{const t=d(n);if(t!=="/")return t;const e=f();return e===""?"/":e},g=n=>l("/apps/"+c+"/?file="+encodeURIComponent(n)),w=(n,t)=>{const e=l("/apps/files/files/"+encodeURIComponent(String(n))),r=new URLSearchParams;return r.set("dir",m(t)),r.set("editing","false"),r.set("openfile","true"),e+"?"+r.toString()},b=(n,t)=>{const e=l("/apps/"+c+"/public/"+encodeURIComponent(n));return t?e+"?file="+encodeURIComponent(t):e},R=()=>{const n=(window.location.pathname||"").match(/(?:\/index\.php)?\/s\/([^/]+)(?:\/.*)?$/);return n&&n[1]||null},U=n=>{if(!n||typeof n!="string")return null;let t;try{t=new URL(n,window.location.origin)}catch{return null}const e=t.pathname||"",r=e.match(/\/apps\/files\/files\/(\d+)\/?$/);if(!r){const a=e.match(/\/f\/(\d+)\/?$/);if(!a){const o=Number(t.searchParams.get("fileid")||"");return Number.isFinite(o)&&o>0?o:null}const i=parseInt(a[1],10);return Number.isFinite(i)&&i>0?i:null}const s=parseInt(r[1],10);return Number.isFinite(s)&&s>0?s:null},I=()=>{const n=(window.location.pathname||"").match(/\/apps\/files\/files\/(\d+)\/?$/);if(!n)return null;const t=parseInt(n[1],10);return Number.isFinite(t)&&t>0?t:null},y=n=>{if(!n||typeof n!="string")return null;let t;try{t=new URL(n,window.location.origin)}catch{return null}const e=(t.pathname||"").match(/(?:\/index\.php)?\/s\/([^/]+)\/download\/?$/);if(!e)return null;const r=(t.searchParams.get("files")||"").trim();if(!p(r))return null;const s=(t.searchParams.get("path")||"/").trim()||"/";return{token:e[1],path:u(s,r)}},x=n=>{if(!n||typeof n!="string")return null;let t;try{t=new URL(n,window.location.origin)}catch{return null}const e=decodeURIComponent(t.pathname||"");if(!e.endsWith(".pad"))return null;const r=["/remote.php/dav/files/","/public.php/dav/files/"].find(o=>e.includes(o));if(!r)return null;const s=e.indexOf(r),a=e.substring(s+r.length),i=a.indexOf("/");return i===-1?null:"/"+a.substring(i+1)};export{I as a,g as b,U as c,y as d,x as e,w as f,f as g,p as i,u as n,R as p,m as r,b as v};
-//# sourceMappingURL=urls-D-89WmZ3.chunk.mjs.map
